mcp_client/requirements.txt
dotenv-python
openai-agents==0.0.11
openai==1.75.0